requirements.txt
address==0.1.1
asgiref==3.2.7
Django==3.0.5
django-cors-headers==3.2.1
django-crispy-forms==1.9.0
djangorestframework==3.11.0
pytz==2019.3
sqlparse==0.3.1
gunicorn==20.0.4
whitenoise==5.2.0

